Refactor: 클래스 이동에 따른 import 수정
Related to: #47
